+++ src/dpapi_ng/_asn1.py
@@ -581,6 +581,22 @@ class ASN1Writer:
         """
         self._data.extend(_pack_asn1_octet_string(value, tag=tag))
 
+    def write_object_identifier(
+        self,
+        value: str,
+        tag: t.Optional[ASN1Tag] = None,
+    ) -> None:
+        """Write an ASN.1 OBJECT_IDENTIFIER value.
+
+        Writes a bytes string value to the current writer.
+
+        Args:
+            value: The object identifier (string representation) to write.
+            tag: Optional tag to use with the value, defaults to the
+                OBJECT_IDENTIFIER universal tag.
+        """
+        self._data.extend(_pack_asn1_object_identifier(value, tag=tag))
+
     def get_data(self) -> bytearray:
         """Gets the data written to the writer.
 
@@ -750,6 +766,35 @@ def _pack_asn1_octet_string(
         tag = ASN1Tag.universal_tag(TypeTagNumber.OCTET_STRING)
 
     return _pack_asn1(tag.tag_class, tag.is_constructed, tag.tag_number, b_data)
+
+
+def _pack_asn1_object_identifier(
+    value: str,
+    tag: t.Optional[ASN1Tag] = None,
+) -> bytes:
+    """Packs an object identifier value represented as string into an
+    ASN.1 OCTET STRING byte value with optional universal tagging."""
+    if not tag:
+        tag = ASN1Tag.universal_tag(TypeTagNumber.OBJECT_IDENTIFIER)
+
+    return _pack_asn1(tag.tag_class, tag.is_constructed, tag.tag_number, _encode_object_identifier(value))
+
+
+def _encode_object_identifier(oid: str) -> bytes:
+    """Encode an object identifier."""
+    cmps = list(map(int, oid.split(".")))
+    if cmps[0] > 39 or cmps[1] > 39:
+        raise ValueError("Illegal object identifier")
+    cmps = [40 * cmps[0] + cmps[1]] + cmps[2:]
+    cmps.reverse()
+    result = []
+    for cmp_data in cmps:
+        result.append(cmp_data & 0x7F)
+        while cmp_data > 0x7F:
+            cmp_data >>= 7
+            result.append(0x80 | (cmp_data & 0x7F))
+    result.reverse()
+    return bytes(result)
 
 
 def _read_asn1_header(